refactor 테스트코드 모듈 분리
