+++ src/test/java/racingcar/module/IntScannerTest.java
@@ -1,10 +1,9 @@
-package racingcar;
+package racingcar.module;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
-import racingcar.module.IntScanner;
 
 public class IntScannerTest {
     void setInputSteam(String input) {
